Add backward-compatibility patch for converters relevant to life support
This patch attempts to upgrade parts that rely on certain obsolete USI
PartModules, to use the new USI_Converter system instead.  It focuses on
modules that are directly needed for life support (habs, recyclers,
greenhouses), and doesn't attempt to handle things like resource
harvesting.  Though it works OK for many parts, there's no guarantee
that it correctly handles all of them, and it is not intended as a
long-term solution; maintainers of mods that integrate with USI should
still update their own integration patches.  This patch is just meant to
get most parts into at least a usable state until then.